rgw: Collapse RGWSI_User_RADOS into RGWSI_User
Preparatory to getting rid of MetaBE.

Also rename datalog_rados to datalog, though it should be moved under
RadosStore, too.
