+++ src/rgw/CMakeLists.txt
@@ -35,7 +35,6 @@ set(librgw_common_srcs
   services/svc_sys_obj_core.cc
   services/svc_tier_rados.cc
   services/svc_user.cc
-  services/svc_user_rados.cc
   services/svc_zone.cc
   services/svc_zone_utils.cc
   rgw_service.cc
