+++ src/rgw/rgw_service.cc
@@ -23,7 +23,7 @@
 #include "services/svc_sys_obj.h"
 #include "services/svc_sys_obj_cache.h"
 #include "services/svc_sys_obj_core.h"
-#include "services/svc_user_rados.h"
+#include "services/svc_user.h"
 
 #include "common/errno.h"
 
@@ -53,7 +53,7 @@ int RGWServices_Def::init(CephContext *cct,
   bi = std::make_unique<RGWSI_BucketIndex>(cct);
   bilog = std::make_unique<RGWSI_BILog>(cct);
   cls = std::make_unique<RGWSI_Cls>(cct);
-  datalog_rados = std::make_unique<RGWDataChangesLog>(cct);
+  datalog = std::make_unique<RGWDataChangesLog>(cct);
   mdlog = std::make_unique<RGWSI_MDLog>(cct, run_sync);
   meta = std::make_unique<RGWSI_Meta>(cct);
   meta_be_sobj = std::make_unique<RGWSI_MetaBackend_SObj>(cct);
@@ -67,7 +67,7 @@ int RGWServices_Def::init(CephContext *cct,
   sync_modules = std::make_unique<RGWSI_SyncModules>(cct);
   sysobj = std::make_unique<RGWSI_SysObj>(cct);
   sysobj_core = std::make_unique<RGWSI_SysObj_Core>(cct);
-  user_rados = std::make_unique<RGWSI_User_RADOS>(cct);
+  user = std::make_unique<RGWSI_User>(cct);
 
   if (have_cache) {
     sysobj_cache = std::make_unique<RGWSI_SysObj_Cache>(dpp, cct);
@@ -75,7 +75,7 @@ int RGWServices_Def::init(CephContext *cct,
 
   vector<RGWSI_MetaBackend *> meta_bes{meta_be_sobj.get(), meta_be_otp.get()};
 
-  bi->init(zone.get(), rados.get(), bilog.get(), datalog_rados.get());
+  bi->init(zone.get(), rados.get(), bilog.get(), datalog.get());
   bilog->init(bi.get());
   bucket_sobj->init(zone.get(), sysobj.get(), sysobj_cache.get(),
                     bi.get(), meta.get(), meta_be_sobj.get(),
@@ -103,8 +103,8 @@ int RGWServices_Def::init(CephContext *cct,
   } else {
     sysobj->init(rados.get(), sysobj_core.get());
   }
-  user_rados->init(rados.get(), zone.get(), sysobj.get(), sysobj_cache.get(),
-                   meta.get(), meta_be_sobj.get(), sync_modules.get());
+  user->init(rados.get(), zone.get(), sysobj.get(), sysobj_cache.get(),
+	     meta.get(), meta_be_sobj.get(), sync_modules.get());
 
   can_shutdown = true;
 
@@ -131,9 +131,9 @@ int RGWServices_Def::init(CephContext *cct,
       return r;
     }
 
-    r = datalog_rados->start(dpp, &zone->get_zone(),
-			     zone->get_zone_params(),
-			     rados->get_rados_handle());
+    r = datalog->start(dpp, &zone->get_zone(),
+		       zone->get_zone_params(),
+		       rados->get_rados_handle());
     if (r < 0) {
       ldpp_dout(dpp, 0) << "ERROR: failed to start datalog_rados service (" << cpp_strerror(-r) << dendl;
       return r;
@@ -215,9 +215,9 @@ int RGWServices_Def::init(CephContext *cct,
       return r;
     }
 
-    r = user_rados->start(y, dpp);
+    r = user->start(y, dpp);
     if (r < 0) {
-      ldpp_dout(dpp, 0) << "ERROR: failed to start user_rados service (" << cpp_strerror(-r) << dendl;
+      ldpp_dout(dpp, 0) << "ERROR: failed to start user service (" << cpp_strerror(-r) << dendl;
       return r;
     }
 
@@ -275,7 +275,7 @@ int RGWServices::do_init(CephContext *_cct, bool have_cache, bool raw, bool run_
   bucket_sync_sobj = _svc.bucket_sync_sobj.get();
   bucket_sync = bucket_sync_sobj;
   cls = _svc.cls.get();
-  datalog_rados = _svc.datalog_rados.get();
+  datalog = _svc.datalog.get();
   mdlog = _svc.mdlog.get();
   meta = _svc.meta.get();
   meta_be_sobj = _svc.meta_be_sobj.get();
@@ -290,7 +290,7 @@ int RGWServices::do_init(CephContext *_cct, bool have_cache, bool raw, bool run_
   sysobj = _svc.sysobj.get();
   cache = _svc.sysobj_cache.get();
   core = _svc.sysobj_core.get();
-  user = _svc.user_rados.get();
+  user = _svc.user.get();
 
   return 0;
 }
@@ -357,7 +357,7 @@ int RGWCtlDef::init(RGWServices& svc, const DoutPrefixProvider *dpp)
   bucket->init(user.get(),
                (RGWBucketMetadataHandler *)bucket_meta_handler,
                (RGWBucketInstanceMetadataHandler *)bi_meta_handler,
-	       svc.datalog_rados,
+	       svc.datalog,
                dpp);
 
   otp->init((RGWOTPMetadataHandler *)meta.otp.get());
